+++ examples/angular/src/server.ts
@@ -14,8 +14,8 @@ import {
 } from 'rxdb/plugins/pouchdb';
 
 // rxdb plugins
-import { RxDBServerPlugin } from 'rxdb/plugins/server';
-addRxPlugin(RxDBServerPlugin);
+import { RxDBServerCouchDBPlugin } from 'rxdb/plugins/server-couchdb';
+addRxPlugin(RxDBServerCouchDBPlugin);
 import { wrappedValidateAjvStorage } from 'rxdb/plugins/validate-ajv';
 
 
@@ -72,7 +72,7 @@ async function run() {
 
 
     // start server
-    const { app, server } = await db.server({
+    const { app, server } = await db.serverCouchDB({
         path: '/' + DATABASE_NAME, // (optional)
         port: COUCHDB_PORT,  // (optional)
         cors: true,   // (optional), enable CORS-headers