+++ src/index.ts
@@ -185,7 +185,7 @@ export type {
     RxQueryOptions,
     RxCouchDBReplicationState,
     RxTypeError,
-    ServerOptions,
+    CouchDBServerOptions,
     SyncOptions,
     SyncOptionsGraphQL,
     MigrationStrategy,