+++ src/types/index.d.ts
@@ -22,7 +22,7 @@ export * from './plugins/replication-graphql';
 export * from './plugins/replication-couchdb';
 export * from './plugins/replication';
 export * from './plugins/local-documents';
-export * from './plugins/server';
+export * from './plugins/server-couchdb';
 export * from './plugins/migration';
 export * from './plugins/backup';
 export * from './plugins/cleanup';