+++ test/unit.test.ts
@@ -23,6 +23,8 @@ import './unit/rx-storage-pouchdb.test';
 import './unit/rx-storage-lokijs.test';
 import './unit/rx-storage-dexie.test';
 
+import './unit/replication-websocket.test';
+import './unit/replication-graphql.test';
 
 
 import './unit/instance-of-check.test';
@@ -54,13 +56,12 @@ import './unit/hooks.test';
 import './unit/orm.test';
 import './unit/replication-protocol.test';
 import './unit/replication.test';
-import './unit/replication-graphql.test';
 import './unit/replication-couchdb.test';
 import './unit/population.test';
 import './unit/leader-election.test';
 import './unit/backup.test';
 import './unit/import-export.test';
-import './unit/server.test';
+import './unit/server-couchdb.test';
 import './unit/dexie-helper.test';
 import './unit/plugin.test';
 import './unit/performance.test';