+++ CHANGELOG.md
@@ -43,6 +43,9 @@
 - CHANGED Attachment data is now always handled as `Blob` because Node.js does support `Blob` since version 18.0.0 so we no longer have to use a `Buffer` but instead can use Blob for browsers and Node.js
 - FIXED multiple problems with encoding attachments data. We now use the `js-base64` library which properly handles utf-8/binary/ascii transformations.
 
+- RENAMED the `server` plugin is now called `server-couchdb` and `RxDatabase.server()` is now `RxDatabase.serverCouchDB()`
+- ADDED the [websocket replication plugin](../replication-websocket.md)
+
 <!-- /CHANGELOG NEWEST -->
 
 <!-- RELEASE BELOW -->